requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0
pydantic==2.6.1
rapidfuzz==3.6.1
pytest==7.4.4
PyYAML==6.0.1
